codec_adapter: rename it to module_adapter and fix all includes
No functional changes.
