+++ src/audio/CMakeLists.txt
@@ -104,8 +104,8 @@ if(NOT CONFIG_LIBRARY)
 	if(CONFIG_COMP_ASRC)
 		add_subdirectory(asrc)
 	endif()
-	if(CONFIG_COMP_CODEC_ADAPTER)
-		add_subdirectory(codec_adapter)
+	if(CONFIG_COMP_MODULE_ADAPTER)
+		add_subdirectory(module_adapter)
 	endif()
 	if(CONFIG_COMP_IGO_NR)
 		add_subdirectory(igo_nr)
