+++ src/audio/module_adapter/CMakeLists.txt
