+++ src/audio/module_adapter/Kconfig
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
-menu "Codec Adapter codecs"
-	visible if COMP_CODEC_ADAPTER
+menu "Module Adapter codecs"
+	visible if COMP_MODULE_ADAPTER
 
 	config CADENCE_CODEC
 		bool "Cadence codec"
