+++ src/audio/module_adapter/module/waves.c
@@ -4,7 +4,7 @@
 //
 // Author: Oleksandr Strelchenko <oleksandr.strelchenko@waves.com>
 //
-#include <sof/audio/codec_adapter/module/generic.h>
+#include <sof/audio/module_adapter/module/generic.h>
 #include <sof/debug/debug.h>
 #include <sof/compiler_attributes.h>
