+++ src/include/sof/audio/module_adapter/module/cadence.h
